requirements.txt
Django
pillow
django-browser-reload
django-cleanup
django-allauth
django-htmx